Factor out CourseCaution component
We will add conditional logic on whether there is a single caution or multiple caution soon, so it's better to factor this out to reduce the component complexity in the course component. Also renamed Caution component to SemesterCaution to disambiguate.
