+++ src/components/Course.vue
@@ -23,15 +23,7 @@
           <span v-if="!compact && semesterString" class="course-semesters">{{
             semesterString
           }}</span>
-          <div v-if="cautionString" class="course-outerWrapper course-tooltip">
-            <div v-if="!compact" class="course-iconWrapper course-iconWrapper--caution">
-              <img class="course-icon course-icon--caution" src="../assets/images/caution.svg" />
-            </div>
-            <div
-              class="course-tooltiptext course-tooltiptext--caution"
-              v-html="cautionString"
-            ></div>
-          </div>
+          <course-caution v-if="cautionString" :compact="compact" :cautionString="cautionString" />
         </div>
       </div>
     </div>
@@ -52,9 +44,11 @@
 <script>
 import Vue from 'vue';
 import CourseMenu from '@/components/Modals/CourseMenu.vue';
+import CourseCaution from '@/components/CourseCaution.vue';
 import { clickOutside } from '@/utilities';
 
 Vue.component('coursemenu', CourseMenu);
+Vue.component('course-caution', CourseCaution);
 
 export default Vue.extend({
   props: {
@@ -306,38 +300,6 @@ export default Vue.extend({
     }
   }
 
-  &-iconWrapper {
-    font-style: normal;
-    display: flex;
-    margin-left: 0.2rem;
-    align-items: center;
-
-    // TODO: styling for info icon on course card
-    // &--info {
-    //   &:before {
-    //     margin-right: 0.2rem;
-    //     font-style: normal;
-    //     content: '|';
-    //   }
-    // }
-
-    &--caution {
-      &:before {
-        margin-right: 0.2rem;
-        font-style: normal;
-        content: '|';
-      }
-    }
-  }
-
-  &-icon {
-    width: 13px;
-
-    &--info {
-      margin-right: 4px;
-    }
-  }
-
   &-buttons {
     display: flex;
     justify-content: space-around;
@@ -361,74 +323,6 @@ export default Vue.extend({
     height: 2.125rem;
     width: 10rem;
   }
-}
-
-// TODO: convert px to rem for spacing
-/* Tooltip container */
-.course-tooltip {
-  position: relative;
-  display: inline-block;
-  // border-bottom: 1px dotted black; /* If you want dots under the hoverable text */
-}
-
-/* Tooltip text */
-.course-tooltip .course-tooltiptext {
-  visibility: hidden;
-  width: 120px;
-  color: $medGray;
-  background-color: $white;
-  text-align: center;
-  padding: 0.5rem;
-  border-radius: 6px;
-  left: -5.2rem;
-  border: 0.75px solid #a7a7a7;
-  top: 1.25rem;
-
-  /* Position the tooltip text */
-  position: absolute;
-  z-index: 1;
-
-  &--caution {
-    width: 7.5rem;
-  }
-}
-
-/* Show the tooltip text when you mouse over the tooltip container */
-.course-tooltip:hover .course-tooltiptext {
-  visibility: visible;
-}
-
-.course-tooltip .course-tooltiptext::after {
-  content: ' ';
-  position: absolute;
-  bottom: 100%; /* At the top of the tooltip */
-  right: 14px;
-  margin-left: -10px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent transparent white transparent;
-  z-index: 3;
-}
-
-.course-tooltip .course-tooltiptext::before {
-  content: ' ';
-  position: absolute;
-  bottom: 100%; /* At the top of the tooltip */
-  right: 12px;
-  margin-left: -2px;
-  border-width: 7px;
-  border-style: solid;
-  border-color: transparent transparent #a7a7a7 transparent;
-
-  z-index: 2;
-}
-
-.course-tooltip .course-tooltiptext--info::after {
-  right: 10px;
-}
-
-.course-tooltip .course-tooltiptext--info::before {
-  right: 8px;
 }
 
 @media only screen and (max-width: 878px) {
